lta-publisher/requirements.txt
functions-framework
requests
google-cloud-pubsub